+++ R/main.R
@@ -9,6 +9,21 @@
 ## Platform: windows (10.0.14393)
 ## ------------------------------------------
 
+#' Check the VS Code Editor
+#'
+#' Check if exists the VS Code Editor.
+#'
+#' @usage exists_vscode()
+#' @return logical class vector that indicates if exists editor VS Code.
+#' @export 
+#' @examples
+#' # default
+#' # exists_vscode()
+
+exists_vscode <- function() {
+    .check_exists_vscode()
+}
+
 #' VS Code Version
 #'
 #' Get the vscode version.
@@ -17,7 +32,8 @@
 #' @return character class vector that indicates the vscode version.
 #' @export 
 #' @examples
-#' vscode_version()
+#' # default
+#' # vscode_version()
 
 vscode_version <- function() {
     .check_vscode_version()
@@ -27,14 +43,18 @@ vscode_version <- function() {
 #'
 #' Get the vscode extension.
 #'
-#' @usage vscode_extension()
+#' @usage vscode_extension(type = "basic", format = TRUE)
+#' @param type character class vector that indicates the extension search type with option 'basic' and 'info'.
+#' @param format logical class vector that indicates if must be the column formated with option 'TRUE' or 'FALSE'.
 #' @return object of the data frame class that gets the vscode extension.
 #' @export 
 #' @examples
-#' vscode_extension()
+#' # default
+#' # vscode_extension()
+#' # vscode_extension(type = "info")
 
-vscode_extension <- function() {
-    .check_vscode_extension()
+vscode_extension <- function(type = "basic", format = TRUE) {
+    .check_vscode_extension(type, format)
 }
 
 #' VS Code Extension Path