+++ NEWS.md
@@ -1,10 +1,25 @@
-# vscoder 0.0.0.9000 (development)
+# vscoder 0.2.2.9000 (2022-07-28)
 
 ## Note from version
 
-### > vscoder (development)
-- development
+### &gt; vscoder (development)
 
-### > vscoder 0.0.0.9000 (2022-07-22)
+-   development
 
-- Start of the package.
+### &gt; vscoder 0.2.2.9000 (2022-07-28)
+
+-   Added function `exists_vscode()` to search installed editor.
+-   `vscode_extension()` received two arguments: type, a character class
+    vector that indicates ‘basic’ or ‘extended’ information about
+    installed extension; and ‘format’, a logical class vector that
+    indicates whether the return of the function should have abbreviated
+    lines.
+-   Internal package improvements.
+    -   The internal error system of the package has been redone for the
+        purpose of maintaining long-term code.
+    -   The package’s internal version control system has been automated
+        for the purpose of maintaining long-term code.
+
+### &gt; vscoder 0.0.0.9000 (2022-07-22)
+
+-   Start of the package.